+++ src/String/StringList.php
@@ -9,8 +9,4 @@ namespace Rowbot\URL\String;
  */
 class StringList extends AbstractStringList implements StringListInterface
 {
-    public function current(): USVStringInterface
-    {
-        return $this->list[$this->cursor];
-    }
 }